Validating new records
Added presence validations for all fields and a format validation for
the email field.

Validations are performed at the "controller" level, or rather at the
create action level. It could be possible to update the core/base_model
object to provide validation support, but given the time constraint,
refactoring that wouldn't be suitable.

Added error messages below each field.

When an error is present, the value sent to the server is kept in the
form when rendering back the response to the submission.

Removed links from the layout that were not used.