Add search input
Added an input to search by city names.

It matches the text case insensitive and also considers partial matches
in the middle or end of the strings.

Each match is enclosed with <mark> html tag.

Added the class .hidden in the safelist of Tailwind, so that it's always
included in the output css.